+++ raiden-dapp/CHANGELOG.md
@@ -7,17 +7,21 @@
 - [#2415] Fix sorting latest transaction on top in transactions list
 - [#2391] Fix notification icon handling for special scenarios
 - [#2410] Fix transfer history list not showing third entry initially
+- [#2426] Display total available capacity in transfer view
 
 ### Added
 
 - [#211] Hub proposal when connecting to new token network
+- [#2435] Sync indicator when connecting to dApp
 - [#2379] Illustrations for info overlays
 
 [#211]: https://github.com/raiden-network/light-client/issues/211
+[#2435]: https://github.com/raiden-network/light-client/issues/2435
 [#2379]: https://github.com/raiden-network/light-client/issues/2379
 [#2415]: https://github.com/raiden-network/light-client/issues/2415
 [#2391]: https://github.com/raiden-network/light-client/issues/2391
 [#2410]: https://github.com/raiden-network/light-client/issues/2410
+[#2426]: https://github.com/raiden-network/light-client/issues/2426
 
 ## [0.14.0] - 2020-11-25